Scatter sky objects before any background sub in dynamicDetection.
Background subtraction makes non-detected annuli around extended
objects so that the footprints of the extended objects will be
kept small. But the non-detected annuli also permit sky objects
to be scattered on the skirts of the extended objects. It should
not happen. One solution is to scatter sky objects before any
background subtractions.